libdevice: fix type of iomux_mux_t in dev_iomux_setup
The iomux mux value type size is configurable but the type in
dev_iomux_setup is fixed to 8 bits.

This commit changes the type of iomux_mux_t to uint32_t so any
configuratin from 8 bits to 32 bits will work.

+++ libdevice/include/device/class/iomux.h
@@ -37,9 +37,9 @@
 
 struct device_iomux_s;
 
-typedef uint8_t iomux_demux_t;
+typedef uint8_t  iomux_demux_t;
 typedef uint16_t iomux_io_id_t;
-typedef uint8_t iomux_mux_t;
+typedef uint32_t iomux_mux_t;
 typedef uint32_t iomux_config_t;
 
 #define IOMUX_INVALID_DEMUX 255
